requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
langchain-chroma
chromadb
gitpython
python-dotenv
watchdog
